requirements.txt
selenium==4.15.2
beautifulsoup4==4.12.2
twilio==8.10.0
APScheduler==3.10.4
python-dotenv==1.0.0
webdriver-manager==4.0.1 